unify coin games code
