+++ src/components/games/coin-3-piles/coin-3-piles.js
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
 import { range } from 'lodash';
-import { strategyGameFactory } from '../strategy-game';
-import { getGameStateAfterMove, getGameStateAfterAiTurn  } from './strategy/strategy';
+import { getGameStateAfterMove  } from './strategy/strategy';
 
-const generateNewBoard = () => ([3, 5, 7]);
-
-const GameBoard = ({ board, setBoard, ctx }) => {
+export const GameBoard = ({ board, setBoard, ctx }) => {
   const [valueOfRemovedCoin, setValueOfRemovedCoin] = useState(null);
   const [hoveredPile, setHoveredPile] = useState(null);
 
@@ -118,36 +115,9 @@ const GameBoard = ({ board, setBoard, ctx }) => {
   );
 };
 
-const getPlayerStepDescription = ({ turnStage }) => {
+export const getPlayerStepDescription = ({ turnStage }) => {
   if (turnStage === 'placeBack') {
     return 'Kattints egy mezőre, hogy visszatégy egy olyan pénzérmét';
   }
   return 'Kattints egy mezőre, hogy elvegyél egy olyan pénzérmét';
-};
-
-const rule = <>
-  Egy kupacban 3 darab 1, 5 darab 2 és 7 darab 3 pengős érme van. Egy lépésben az
-  éppen soron lévő játékos elvesz egy érmét a kupacból, és helyette berakhat egy darab kisebb
-  értékű érmét, vagy dönthet úgy, hogy nem tesz be semmit. Az nyer, aki elveszi az utolsó érmét
-  a kupacból.
-
-  Te döntheted el, hogy kezdeni szeretnél-e, vagy második játékos lenni.
-  Sok sikert! :)
-</>;
-
-const Game = strategyGameFactory({
-  rule,
-  title: '3 db 1, 5 db 2 és 7 db 3 pengős',
-  GameBoard,
-  G: {
-    getPlayerStepDescription,
-    generateNewBoard,
-    getGameStateAfterAiTurn
-  }
-});
-
-export const Coin357 = () => {
-  const [board, setBoard] = useState(generateNewBoard());
-
-  return <Game board={board} setBoard={setBoard} />;
 };
