+++ src/components/games/coin-3-piles/strategy/strategy.spec.js
